Refactor CommandEvents interface and integrate command decoration methods into RecordConsumerSupport
